accessing ingredients from quiz component state and sorted seed data array in order
